+++ financeiro/admin/saida_admin.py
@@ -11,3 +11,27 @@ class SaidaAdmin(admin.ModelAdmin):
     "observacoes",
     "usuario_criacao",
   ]
+
+  search_fields= [
+    "id",
+    "valor",
+    "dataCriacao",
+    "motivo",
+    "observacoes",
+    "usuario_criacao__username",
+  ]
+
+  list_filter= [
+   "dataCriacao",
+   "usuario_criacao", 
+  ]
+
+  autocomplete_fields= [
+    "usuario_criacao",
+  ]
+
+  fields= [
+    "valor",
+    "motivo",
+    "observacoes",
+  ]
