+++ reservas/admin/__init__.py
@@ -1,0 +1,7 @@
+from .quarto_admin import QuartoAdmin
+from .reserva_admin import ReservaAdmin
+
+__all__ = [
+  QuartoAdmin,
+  ReservaAdmin,
+]
