+++ reservas/models/__init__.py
@@ -1,0 +1,7 @@
+from .quarto import Quarto
+from .reserva import Reserva
+
+__all__ = [
+  Quarto,
+  Reserva,
+]
